seperate notifier into its own file
